Sneakily store strategy names in the docstring
I like this better because you don't have to maintain your list of
strategies in two places; instead, you just document the name of the
strategy in the docstring. Of course this has a tradeoff: everything you
put in the docstring is now the implicit name of the strategy. If you
want longer docstrings, this could be easily fixed with a function
`doc(strategy)` that might take a strategy and pull only the first line
from the docstring.