Consistent formatting
The formatting in these two files was particularly inconsistent, with a
mix of tabs and spaces and irregular indenting. I basically ran it
through a program that auto-indented everything instead of doing it
manually, so there may be a few things that are off about it, but the
formatting is generally consistent now.