Extract common calls into a function
I actually think that a lot of this could be better handled by a data
structure that represents all of the lines and their associated data.
Right now we have to have these global variables and then remember about
each specific strategy. Instead, we should have a "strategies" object
that contains all of this information, and then do generic operations on
it.

I'll address this in the next commit :)